requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.3.0
Pillow>=9.0.0
qrcode[pil]>=8.0.0